% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCellTypes.R
\name{getCellTypes}
\alias{getCellTypes}
\title{Get the cell-type content for each sample from the panglaoDB database.}
\usage{
getCellTypes(
  sra = "All",
  srs = "All",
  tissue = "All",
  protocol = "All",
  specie = "All",
  verbose = TRUE
)
}
\arguments{
\item{sra}{A}

\item{srs}{A}

\item{tissue}{A}

\item{protocol}{A}

\item{specie}{A}

\item{verbose}{A}
}
\description{
Get the cell-type content for each sample from the panglaoDB database.
}
