% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSamples.R
\name{getSamples}
\alias{getSamples}
\title{Download the expression matrix and annotations from the panglaoDB database.}
\usage{
getSamples(srs)
}
\arguments{
\item{srs}{A}
}
\description{
Download the expression matrix and annotations from the panglaoDB database.
}
