% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMarkers.R
\name{getMarkers}
\alias{getMarkers}
\title{Get the list of samples from the panglaoDB database containing a set of molecular markers.}
\usage{
getMarkers(markerGenes)
}
\arguments{
\item{markerGenes}{A}
}
\description{
Get the list of samples from the panglaoDB database containing a set of molecular markers.
}
