% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSampleComposition.R
\name{getSampleComposition}
\alias{getSampleComposition}
\title{Get the cell-type content for each sample from the panglaoDB database.}
\usage{
getSampleComposition(
  sra = "All",
  srs = "All",
  tissue = "All",
  protocol = "All",
  specie = "All",
  verbose = TRUE
)
}
\arguments{
\item{sra}{The SRA identifier of the biological sample in the SRA database}

\item{srs}{The SRS identifier of the biological sample in the SRA database}

\item{tissue}{The tissue from which the biological samples originates from}

\item{protocol}{The single-cell library preparation protocol used to generate the data}

\item{specie}{The specie from which the biological samples originates from. The corresponding reference genome was used for maping the sequencing reads}

\item{verbose}{A boolean value TRUE or FALSE to activate the verbose mode}
}
\description{
Get the cell-type content for each sample from the panglaoDB database.
}
\examples{
# From PanglaoDB
# https://panglaodb.se/list_clusters_and_cell_types.html?sra=SRA689041&srs=SRS3166675

sc_SRS3166675 <- getSampleComposition(srs = 'SRS3166675')
sc_SRS3166675

}
