% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSampleList.R
\name{getSampleList}
\alias{getSampleList}
\title{Get the sample list from the panglaoDB database.}
\usage{
getSampleList()
}
\description{
Get the sample list from the panglaoDB database.
}
\examples{
sampleList <- getSampleList()
head(sampleList)
}
